+++ src/pages/begin.tsx
@@ -13,6 +13,7 @@ export default function Begin() {
   const handleSubmit = (e: React.ChangeEvent<HTMLFormElement>) => {
     e.preventDefault();
     // API 전송
+    console.log(nickname);
     if (nickname === '') {
       setModal(true);
     }
@@ -47,9 +48,7 @@ export default function Begin() {
             </span>
           )}
         </div>
-        <div
-          className={`fixed inset-x-0 bottom-[2.8125rem]  m-auto w-[25.625rem]`}
-        >
+        <div className={`inset-x-0 m-auto mt-44 w-full`}>
           <Button
             disabled={!reg.test(nickname)}
             onClick={() => {
